Revert "new structure"
This reverts commit e0943ad
