--- app/main/__init__.py
@@ -1,20 +1,0 @@
-# coding=utf8
-# Copyright (c) 2018 CineUse
-
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_bcrypt import Bcrypt
-
-from .config import config_by_name
-
-db = SQLAlchemy()
-flask_bcrypt = Bcrypt()
-
-
-def create_app(config_name):
-    app = Flask(__name__)
-    app.config.from_object(config_by_name[config_name])
-    db.init_app(app)
-    flask_bcrypt.init_app(app)
-
-    return app
