--- app/main/config.py
@@ -1,44 +1,0 @@
-# coding=utf8
-# Copyright (c) 2019 CineUse
-import os
-
-# 下面一行用于指定数据库url
-# postgres_local_base = os.environ['DATABASE_URL']
-
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-
-class Config:
-    SECRET_KEY = os.getenv('SECRET_KEY', 'my_precious_secret_key')
-    DEBUG = False
-
-
-class DevelopmentConfig(Config):
-    # 下面一行用于postgres类型数据库
-    # SQLALCHEMY_DATABASE_URI = postgres_local_base
-    DEBUG = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'flask_boilerplate_main.db')
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-
-class TestingConfig(Config):
-    DEBUG = True
-    TESTING = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'flask_boilerplate_test.db')
-    PRESERVE_CONTEXT_ON_EXCEPTION = False
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-
-class ProductionConfig(Config):
-    DEBUG = False
-    # 下面一行用于postgres类型数据库
-    # SQLALCHEMY_DATABASE_URI = postgres_local_base
-
-
-config_by_name = dict(
-    dev=DevelopmentConfig,
-    test=TestingConfig,
-    prod=ProductionConfig
-)
-
-key = Config.SECRET_KEY
