--- app/main/util/__init__.py
@@ -1,2 +1,0 @@
-# coding=utf8
-# Copyright (c) 2019 CineUse
