--- app/main/util/dto.py
@@ -1,18 +1,0 @@
-# coding=utf8
-# Copyright (c) 2019 CineUse
-
-"""
-Data Transfer Object is responsible for carrying data between processes.
-"""
-
-from flask_restplus import Namespace, fields
-
-
-class UserDto:
-    api = Namespace('user', description='user related operations')
-    user = api.model('user', {
-        'email': fields.String(required=True, description='user email address'),
-        'username': fields.String(required=True, description='user username'),
-        'password': fields.String(required=True, description='user password'),
-        'public_id': fields.String(description='user Identifier')
-    })
