--- app/test/test_config.py
@@ -1,51 +1,0 @@
-# coding=utf8
-# Copyright (c) 2019 CineUse
-
-import os
-import unittest
-
-from flask import current_app
-from flask_testing import TestCase
-
-from app.manage import app
-from app.main.config import basedir
-
-
-class TestDevelopmentConfig(TestCase):
-    def create_app(self):
-        app.config.from_object('main.config.DevelopmentConfig')
-        return app
-
-    def test_app_is_development(self):
-        self.assertFalse(app.config['SECRET_KEY'] is 'my_precious')
-        self.assertTrue(app.config['DEBUG'] is True)
-        self.assertFalse(current_app is None)
-        self.assertTrue(
-            app.config['SQLALCHEMY_DATABASE_URI'] == 'sqlite:///' + os.path.join(basedir, 'flask_boilerplate_main.db')
-        )
-
-
-class TestTestingConfig(TestCase):
-    def create_app(self):
-        app.config.from_object('main.config.TestingConfig')
-        return app
-
-    def test_app_is_testing(self):
-        self.assertFalse(app.config['SECRET_KEY'] is 'my_precious')
-        self.assertTrue(app.config['DEBUG'])
-        self.assertTrue(
-            app.config['SQLALCHEMY_DATABASE_URI'] == 'sqlite:///' + os.path.join(basedir, 'flask_boilerplate_test.db')
-        )
-
-
-class TestProductionConfig(TestCase):
-    def create_app(self):
-        app.config.from_object('main.config.ProductionConfig')
-        return app
-
-    def test_app_is_production(self):
-        self.assertTrue(app.config['DEBUG'] is False)
-
-
-if __name__ == '__main__':
-    unittest.main()
