+++ passenger_wsgi.py
@@ -1,1 +1,5 @@
-application = StrackGitServerHooks.app
+import os
+import sys
+import app
+
+application = app.app
